refactor: split worktree and connection-settings services by responsibility
WorktreeService keeps the session worktree policy (prepare/diff/merge/
discard/recover) while the stateless git shell layer moves to
worktree-git.ts: runGit, gitRoot, currentBranch, worktreeDirty,
defaultBranch, path helpers and the .dsh-worktrees exclude entry.

ConnectionSettingsService keeps state + lifecycle; the pure wire mapping
(normalize, relay profile/model shapes, provider views, valueAt/valueOf)
moves to connection-settings/mapping.ts and the four idempotent
connect-time migrations to connection-settings/migrations.ts. Public
exports (constants, types, service) are unchanged.
